crawler/requirements.txt
# Core dependencies
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiohttp==3.9.1
python-dotenv==1.0.0
pdfplumber==0.10.3
pandas==2.1.3
python-docx==1.0.1
rich==13.7.0

# Optional S3 support
boto3==1.29.3

# Development
pytest==7.4.3
black==23.11.0
